+++ packages/react/src/ErrorBoundary.test-d.tsx
@@ -42,7 +42,7 @@ describe('<ErrorBoundary/>', () => {
   })
 
   it('type check', () => {
-    expectTypeOf(<ErrorBoundary.Consumer>{() => <></>}</ErrorBoundary.Consumer>).toEqualTypeOf<JSX.Element>()
+    expectTypeOf(<ErrorBoundary.Consumer>{() => <></>}</ErrorBoundary.Consumer>).toEqualTypeOf<React.JSX.Element>()
     expectTypeOf(<ErrorBoundary.Consumer>{() => <></>}</ErrorBoundary.Consumer>).not.toEqualTypeOf<ReactNode>()
   })
 })