+++ packages/react/src/ErrorBoundaryGroup.test-d.tsx
@@ -3,7 +3,9 @@ import { ErrorBoundaryGroup } from './ErrorBoundaryGroup'
 
 describe('ErrorBoundaryGroup', () => {
   it('type check', () => {
-    expectTypeOf(<ErrorBoundaryGroup.Consumer>{() => <></>}</ErrorBoundaryGroup.Consumer>).toEqualTypeOf<JSX.Element>()
+    expectTypeOf(
+      <ErrorBoundaryGroup.Consumer>{() => <></>}</ErrorBoundaryGroup.Consumer>
+    ).toEqualTypeOf<React.JSX.Element>()
     expectTypeOf(
       <ErrorBoundaryGroup.Consumer>{() => <></>}</ErrorBoundaryGroup.Consumer>
     ).not.toEqualTypeOf<ReactNode>()