+++ packages/react/src/hooks/useIsomorphicLayoutEffect.ts
@@ -1,2 +1,0 @@
-import { useEffect, useLayoutEffect } from 'react'
-export const useIsomorphicLayoutEffect = typeof window !== 'undefined' ? useLayoutEffect : useEffect